Accept advisories with respin suffixes. BZ 1065380
We don't typically publish respin numbers for advisories in updateinfo*
and admins are never supposed to reference them when updating their
systems.  However, we do display them in some places like the RHN portal
[1], even for advisories that were never respun.  For this reason, we
should accept such extended IDs as well and treat them as if they were
specified without the suffix.

* There seem to have been such advisories in the past, see [2].  To have
them covered, let's also match the exact IDs as specified on the
cmdline.

For example, this command:
  yum updateinfo RHSA-2016:1234-1
will now match any advisories that have RHSA-2016:1234-1 or
RHSA-2016:1234 as their ID in updateinfo.

Note that this change affects all the places on the cmdline where
advisory IDs are accepted and only applies to Red Hat errata.

[1] https://rhn.redhat.com/errata/RHBA-2010-0943.html
[2] http://www.cyberciti.biz/faq/redhat-fedora-centos-linux-yum-installs
    -security-updates/